Move create_action from category_controller to urls_controller
